third_party/blink/web_tests/http/tests/security/xssAuditor/html5-import-CORS-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3clink%20rel=%22import%22%20href=%22http://127.0.0.1:8000/security/xssAuditor/resources/xss.js%22%3e' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
CONSOLE WARNING: HTML Imports is deprecated and will be removed in M80, around February 2020. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
This test passes if the XSSAuditor blocks the load prior to the CORS restriction kicking in. We've not bothered to enable CORS for this test, unlike what a real attacker would do, so a CORS error here means failure.
